invest_telegram/requirements.txt
telethon>=1.36.0
anthropic>=0.40.0
python-dotenv>=1.0.0
schedule>=1.2.0
pykrx>=1.0.45
matplotlib>=3.8.0
